Remove core page variables and dataLayer patch for UA/GA3 Fix #10733
Removes some code that supported our weird UA/GA3 setup including: page id, core page, dataLayer over-rides for click tracking, and variables indicating if a page has a video and download button.

- Remove: core-dataLayer.js and associated tests & init
- Functionality removed:
  - page-id
  - dataLayer "monkey patch" for UA
  - pageHasDownload, PageVersion, and releaseWindowVersion variables
- Removed jinja template code that supported the page-id JS
- Moved: amoExperiments to new files, converted to es6
