+++ bedrock/base/templates/404.html
@@ -6,8 +6,6 @@
 
 {% extends "base-error.html" %}
 
-{% block gtm_page_id %}data-gtm-page-id="404"{% endblock %}
-
 {% block page_title %}{{ ftl('not-found-page-not-found-page-page-not-found') }}{% endblock %}
 
 {% block page_css %}
