+++ bedrock/base/templates/500.html
@@ -6,8 +6,6 @@
 
 {% extends "base-error.html" %}
 
-{% block gtm_page_id %}data-gtm-page-id="500"{% endblock %}
-
 {% block page_title %}{{ ftl('error-page-error-page-internal-server-error') }}{% endblock %}
 
 {% block page_css %}
