+++ bedrock/firefox/templates/firefox/releases/dev-browser-notes.html
@@ -6,7 +6,6 @@
 
 {% extends "firefox/releases/notes.html" %}
 
-{% set page_id = 'data-gtm-page-id="/firefox/auroranotes/"' %}
 {% set product_name = 'Firefox Developer Edition' %}
 {% set header_class = 'mzp-t-dark t-developer' %}
 {% set brand_product = 'developer' %}
