+++ bedrock/firefox/templates/firefox/releases/nightly-notes.html
@@ -21,7 +21,6 @@
 
 {% set header_class = 'mzp-t-dark t-nightly' %}
 {% set brand_product = 'nightly' %}
-{% set page_id = 'data-gtm-page-id="/firefox/nightly/releasenotes"' %}
 
 {% if release.is_latest %}
   {% set primary_heading = 'See what landed recently in <span>Firefox Nightly</span>!'|safe %}
