+++ bedrock/firefox/templates/firefox/releases/release-notes.html
@@ -8,7 +8,6 @@
 {% from "macros.html" import google_play_button, apple_app_store_button with context %}
 
 {% if release.product == 'Firefox for iOS' %}
-  {% set page_id = 'data-gtm-page-id="/firefox/ios/releasenotes/"' %}
   {% set product_name = release.product %}
   {% set download_button_primary %}
     {{ apple_app_store_button(href=app_store_url('firefox', 'firefox-release-notes'), id='download-ios-primary') }}
@@ -17,7 +16,6 @@
     {{ apple_app_store_button(href=app_store_url('firefox', 'firefox-release-notes'), id='download-ios-secondary') }}
   {% endset %}
 {% elif release.product == 'Firefox for Android' %}
-  {% set page_id = 'data-gtm-page-id="/firefox/android/releasenotes/"' %}
   {% set product_name = 'Firefox for Android' %}
   {% set download_button_primary %}
     {{ google_play_button(id='download-android-primary') }}
