+++ bedrock/firefox/templates/firefox/whatsnew/base.html
@@ -9,8 +9,6 @@
 {# "noindex" pages should not have the canonical or hreflang tags: bug 1442331 #}
 {% block canonical_urls %}<meta name="robots" content="noindex,follow">{% endblock %}
 
-{% block gtm_page_id %}data-gtm-page-id="/firefox/whatsnew/"{% endblock %}
-
 {# Exclude stub attribution for in-product pages: issus 9620 #}
 {% block stub_attribution %}{% endblock %}
