+++ bedrock/mozorg/templates/mozorg/home/base.html
@@ -6,8 +6,6 @@
 
 {% extends "base-protocol-mozilla.html" %}
 
-{% block gtm_page_id %}data-gtm-page-id="Homepage"{% endblock %}
-
 {% block extra_meta %}
 <!-- validates bing webmaster tools -->
 <meta name="msvalidate.01" content="B7B177115A634927D608514DA17B2574">
