+++ bedrock/newsletter/templates/newsletter/country.html
@@ -6,8 +6,6 @@
 
 {% extends 'newsletter/base.html' %}
 
-{% block gtm_page_id %}data-gtm-page-id="/newsletter/country/"{% endblock %}
-
 {# "noindex" pages should not have the canonical or hreflang tags: bug 1442331 #}
 {% block canonical_urls %}<meta name="robots" content="noindex,follow">{% endblock %}
