+++ bedrock/newsletter/templates/newsletter/management.html
@@ -8,8 +8,6 @@
 
 {# Template used for a user to manage their subscriptions #}
 
-{% block gtm_page_id %}data-gtm-page-id="/newsletter/existing/"{% endblock %}
-
 {# "noindex" pages should not have the canonical or hreflang tags: bug 1442331 #}
 {% block canonical_urls %}<meta name="robots" content="noindex,follow">{% endblock %}
